refactor: restructure project folders
