+++ React/panda-market/src/components/Layout/Header.jsx
@@ -1,6 +1,6 @@
 import { Link, NavLink } from "react-router-dom";
-import logoImg from "../assets/logo/panda-market-logo.svg";
-import UserIcon from "../assets/user/default-profile.png";
+import logoImg from "../../assets/logo/panda-market-logo.svg";
+import UserIcon from "../../assets/user/default-profile.png";
 import styles from "./Header.module.css";
 
 function getLinkStyle({ isActive }) {
