+++ React/panda-market/src/pages/MarketPage/components/ItemList.jsx
@@ -1,4 +1,3 @@
-import Container from "../../../components/Container";
 import ItemCard from "./ItemCard";
 import styles from "./ItemList.module.css";
 
@@ -6,7 +5,7 @@ function ItemList({ items, best }) {
   const ClassNames = `${best ? styles.bestList : ""} ${styles.itemList}`;
 
   return (
-    <Container>
+    <>
       <ul className={ClassNames}>
         {items.map((item) => (
           <li key={item.id}>
@@ -14,7 +13,7 @@ function ItemList({ items, best }) {
           </li>
         ))}
       </ul>
-    </Container>
+    </>
   );
 }
 export default ItemList;
